Added google analytics to base ui template
- relevant tickets:

closes #135

- what's this do?

this adds google analytics to the base template for the ui app.

first, this adds a template tag for the ui app that lets us pull values
out of `settings.py` (we store the google analytics id there).

then we can just pop the google analytics JS snippet into the `<head>`
of our base template, and pull in the analytics id using the template
tag. Nice!

- where should the reviewer start?

`/ui/templatetags/ui_extras.py` <- this is the thing I'm least sure
about.

- how should this be manually tested?

we can generate a throwaway google analytics account, set the
environment variable `GA_TRACKING_ID` to the tracking id for that
account (in the micromasters `.env`), and then confirm that visiting any
page results in a page view being registered.

- any background context?

I was a little unsure as to how to get the environment variable defining
our tracking ID into the template. this SE thread was handy:

https://stackoverflow.com/questions/433162/can-i-access-constants-in-settings-py-from-templates-in-django