Updated eslint configuration
This is now pulling in the eslint recommended settings and the
eslint-react recommended settings (which automatically enables support
for proper interpretation of JSX).

there are several errors that come up with this config - it's definitely
more aggressive than what we were doing before.

not sure if we want this much linting on JS - I personally like linters
because then code is very predicated, but then we all have to agree on
things like ' vs ", dangling commas, semicolons, etc.