Added react-router to profile tabs, moved redux form logic to shared module
here we switch to using react-router to navigate between the profile
tabs.

this comprises a couple of changes:

- route changes (routes under /profile for the tabs)
- slight refactor of ProfilePage to work with this.props.children,
  instead of explicit calls to components
- pull the logic for creating form fields and callbacks (to update the
  store) into a shared file, so they may be used on different tabs

so now the tabs on the `/profile` page navigate between tabs and URLs.

the tabs are dynamically generated based on the paths that are children
of `/profile` (so, right now, 'personal' and 'professional'). this is
kind of nice - we can just add another route to `dashboard.js` under the
`/profile` route and it will automatically have a correctly wired up Tab
in the view, with a correcty `onChange` callback and so on.

this accomplishes part of #150.