Deleted the `employment` route placeholder
This removes the employment route placeholder from our router. We
probably aren't going to have the employment tab ready in time for the
demo, and we'd rather not show off the placeholder.

    modified:   static/js/dashboard.js

This just removes the route for the `EmploymentTab` component.

    modified:   static/js/components/PersonalTab.js

This sets the `PersonalTab` component to redirect to `/dashboard` on
save, instead of to `/profile/professional`.