Added ProfileTab utility class
- What are the relevant tickets?

none

- What's this PR do?

this creates a utility class called `ProfileTab` which lets us cut down
on boilerplate a little (when setting up the profile tabs).

we just write components that inherit from `ProfileTab`, and the
`boundTextField` and so on functions are already bound to `this` (just
remember to call `super(props)`in your constructor!).

- Where should the reviewer start?

read the new class (`static/js/util/ProfileTab.js) and review the
changes to the personal tab (`static/js/components/PersonalTab.js`).

- How should this be manually tested?

Make sure this doesn't clobber any existing behavior.

- Any background context you want to provide?

my name is alice and I hate boilerplate.