Added confirmation dialog when deleting education and work entries
This adds a confirmation dialog that pops-up when clicking the 'delete'
icon for a work or education history entry. This entails some associated
reducers and actions, and a new class called `ConfirmDeletion` which
will, hopefully, be generally useful for this sort of thing.

Most of the key stuff ends up being implemented as instance methods on
`ProfileFormFields`.