Added check to only show edit buttons for a user's own profile
This adds a little utility function called `userPrivilegeCheck`, which
basically just checks that the username on the profile in question is
the same as `SETTINGS.username`. We pass a callback in to
`userPrivilegeCheck`, and `userPrivilegeCheck` returns the value of
executing that callback if the usernames match, else it returns an
optional second argument or, if that isn't provided, it returns
undefined.