Condensed EducationForm on `/users`
This adds a couple new components to enable displaying a more condensed
form of the education history on the /users page. We want to show
something a little more compact and polished there, and we want it to
look like a nice presentational component when viewing another user's
profile.

We retain on `/users` the ability to add, delete, and edit education
history entries.