Added 'confirm delete all entries' when closing switches
This changes the behavior of the switches that we display on the
education and work history profile steps. Now it looks like:

- if the user has not entered anything she can freely toggle the switch
- if there is an entry saved, clicking the switch will display a popup
- confirming the popup will delete all entries and toggle the switch to
  'off'.
- cancelling the popup will close the dialog, leave the entries and
  current switch position intact.

On the work history step this applies to all work history entries, on
the education step we do it level-by-level, so toggling and confirming
for a particular level will delete only the entries for that level.