Removed routing from profile flow
This removes routing from the initial profile flow, so that instead of
guiding the user through a series of components which each live at a
different url (`/profile/education` and so on) everything happens at a
single url (`/profile`).

We're now managing the `profileStep` with a redux action. This entails a
pretty big set of changes to `ProfilePage` and `ProfileFormContainer`,
and a smaller set of changes to `SettingsPage` and some other files.

This also changes `validateProfileComplete`, since after determining the
profile is incomplete we now want to dispatch a particular action and
redirect to `/profile`, instead of redirecting to
`/profile/professional`.

`ProfileProgressControls` now accepts two callbacks (or `void`) to
attach to `onClick` on it's buttons, instead of URLs to push onto
`this.context.router`.

Various other changes are necessary to deal with breaking tests and so
on.