Updated validation state when editing fields
This changes our profileEdit functions so that, given a validator
function, changing or updating a field in such a way that a validation
error is fixed will cause the error to go away.

This is accomplished by checking for existing error messages, and, given
some existing validation errors, re-checking the profile and updating
the store with the result. This means that we can fix errors in-line as
the user is typing or selecting, but we do not show validation errors
before any input has been entered.

this closes #549