Refactored ProfileFormContainer to remove boilerplate
This adds a few functions in `util/redux.js` which make it easier to
generate action helpers. Action helpers are functions that will take an
argument and dispatch a change to the store - we use them to pass down
callbacksfrom container components to presentational components as
props.

This lets us avoid a lot of boilerplate we had on `ProfileFormContainer`
which was basically just pulling `dispatch` out of `this.props` and then
dispatching a particular action.

closes #712