Implemented new navbar design
This implements the updated Navbar design. It's full-width, fixed to the
top of the page, and has highlighted tabs that show which link in the
nav menu you're on via highlighting and underlining. The navbar also now
lives in Navbar.js, which was previously called Header.js - I'm using a
react-mdl component called Header, so I renamed the component to avoid
the naming collision.

closes #723