Implemented new design for user menu
This updates the user menu (what we show in the top right of the page)
to reflect the latest designs. Before we were using a bootstrap
component, now we're using just React + CSS to render the user menu,
with a small react dependency added to enable 'click outside to close'
behavior.

closes #724