Added 'jump to error' on profile forms
This adds a nice little behavior that should make it more obvious to
users with short screens where validation errors are coming from.
Previously it was possible for a user with a short screen height to be
unable to see all the fields in a form dialog. It's possible in that
scenario for the form to have a validation error in a field which the
user cannot see, leaving them with no indication of why the thing won't
save.

This adds a pretty simple little thing: when we're updating the
validation, we set a selector on all form inputs that have validation
errors. Then we use `document.querySelector` to grab the first one of
these, and we call `.scrollIntoView` on it. Nice! This way it's pretty
hard for a user to miss the invalid field.

closes #561