Implemented new profile form design
This implements the latest design for the profile forms. This
necessitates a lot of changes to the various form components, and also a
change to the overall structure of how we render the app. It's now
something like this:

```html
<div id="app">
  <div class="micromasters-navbar">
    <!-- header contents -->
  </div>
  <div class="page-contents">
    <!-- the UI for whatever page (top-level container) we're on -->
  </div>
  <footer>
    <!-- the page footer -->
  </footer>
</div>
```

which should make things in general a lot easier to style and lay out in
the future.

closes #769