Removed .name call on anonymous functions for createActionHelper
This fixes a bug that showed up when we built the JS for release. The
minifier was changing the names of the variables that a number of
anonymous functions were bound to. I had written some code that depended
on the `.name` attribute of those functions, which was not the same
after minification.

Anway, this fixed it!

closes #788