installed flow v0.30.0
This upgrades flow to v0.30. This entailed a bunch of trivial changes to
some type signatures (React$Element -> React$Element<*>, for example)
and some more substantial changes to a couple of object types.

Also this installed the Sanctuary and Ramda functional programming
libraries, which we're using initially in the date validation and
boundDateField method (on profile_edit.js), which has greatly simplified
some of the null / undefined checking we had before. Additionally, there
are some utility functions for working with Sanctuary types in
`util/sanctuary.js`.

Also we're now using flow-typed to provide library definitions for flow.
This PR checks in definitions for redux and moment.js, and there are
many other definitions available for popular npm packages.

closes #784