Added signup dialog to homepage and program pages
This adds a little signup dialog to the homepage and the program pages.
This lets the user select a program to enroll in, from the list of all
live programs. On program pages the program corresponding to the page is
pre-selected.

Program selection in the dialog is managed via a redux store, which is
synced with localStorage. This means that when the user comes back from
signing in to edX we can sync up the same reducer with localStorage, and
we can then issue a request to our program enrollment API to add a new
enrollment.

This also refactors the store configuration functions (in
`configureStore.js`) and adds a polyfill / test mock for localStorage
for use in unit tests.

closes #917