Added default currency (based on country) to calculator
Here we add some code to set a default currency when the user opens the
financial aid calculator. We take their country (off of their profile)
and use that to try to get the usual currency for that country.

This also adds some code to exclude currencies from the option list
which are not available on the backend.

closes #1048
closes #1282