Fixed user menu wideness
The user menu div width was based on the width of it's parent, which
depended on the length of the user's preferred name. Thus, if a user had
a very long preferred name, the user menu could become very wide in a
silly way. This just sets a `max-width` on the div to avoid that.

closes #1338