Implemented mobile sidenav
This implements a mobile side navigation drawer. We show a hamburger menu
to open the nav drawer if the user's device is less than 900px wide. The
user can navigate around the site using the links in the drawer, can
edit their photo, or, on certain pages, change their current program.

The user can click outside or swipe to dismiss the drawer.

closes #1414