Note Flow incompatibility within Docker (#1469)
* Note Flow incompatibility within Docker

I believe this is because Flow runs a binary file. If the
host OS is not Linux, npm will download and install the Flow binary
that is compatible with your host OS, not the one that is
compatible with Linux on the Docker image. Since our Docker images
are set up to mount the local filesystem inside the image,
the Docker image picks up the Flow binary for the host OS, instead
of downloading and installing the proper Flow binary for the
Docker OS.

* Note that docker-compose works for Linux folks