Refactored date validation
Here I refactored our date validation code to fix a couple of edge
cases, and for general code cleanliness and re-use. This also adds
validation for the case where an employment end date is in the future,
and added the capability to have an education end date up to 10 years in
the future.

closes #1462