Prevented the user from issuing multiple image upload requests
Currently on master it is possible to issue multiple concurrent image
upload requests, by 'cancelling' the first and then opening another.  On
a slow connection this could compound the existing issues with image
uploads. This change will selectively persist the image edit state, when
there is a request currently in-flight. This means that if the user
cancels the first dialog and re-opens it they will see a spinner
indicating that the first request is still in-flight, and they will be
prevented from issuing another one.

closes #2020
pr #2039