Updated app to proxy requests through Nginx (#2063)
* Updated app to proxy requests through Nginx

Proxying apps through Nginx solves the chunked-content issues that we
have been seeing with the profile image uploads. This commit updates
the docker-compose and procfiles to use Nginx in front of uWSGI for
serving all routes.

* Explicitly added uwsgi params to nginx config

The heroku buildpack doesn't include the uwsgi_params file so I added
the settings that it includes. Also addressed current review comments.

* Added staticfiles to asset paths

Added the staticfiles directory to the list of paths searched for
static assets.

* Added missing webpack location to try_files

Added the `@webpack` named location to the `try_files` directive for
static assets.