Removed foo: Function = () => style annotations
We had a lot of arrow functions declared on classes, annotated with the
Function type, like this:

```
class Foo {
  bar: Function = () => (
    'potatoes'
  );
}
```

Turns out that annotating functions like that is sort of equivalent to
annotating them with the any type, meaning that the return types and so
on aren't checked. Not good!

We had to annotate class properties like this at one point, because flow
did now have support for inferring the types of function literals (arrow
functions). But, now it does! So we can remove this, and our type
checking will benefit.

pr #2145