Updated Webpack version, split out LearnerSearchPage code
This updates Webpack to 2.2.0-rc.3 and implements everything we'll need
for react-router route based splitting. For now we're still importing
all of the relevant components in the dashboard_routes file except
for LearnerSearchPage, which means that this is the only thing we're
bundling separately. This should prevent normal users from
downloading all of Searchkit.

Later we'll follow-on with a PR to add a spinner in-between page loads,
and we'll turn on route splitting everywhere once we do that.

another part of #2163
pr #2247