Fixed a bug on the profile with setting the program
This fixes a small UI bug with the program selector that we show on
/profile/personal.

Basically, if a user returns to /profile/personal to change something,
and they have already enrolled in a program, we don't want the program
selector to be empty. If the program selector is empty, when the user
presses 'save' we will show a validation error - if the user is already
enrolled in a program this is inappropriate.

I made a change in 72504a58efe6d679e3e3f69f7d76e5a67a83e2f4 that was
supposed to address this. My earlier fix made sure that, after we fetch
the program enrollments, we check if the user is on /profile and, if so,
we dispatch the action to pre-set the user's first enrolled program,
which would prevent the validation error from showing up.

There is a bug in this fix, though. In the case that the user had a
validation error which would cause them to be redirected from /dashboard
-> /profile, the request to fetch enrollment information could come back
while the user was still on the dashboard. Then the pathname check would
fail, and so the user would still see the validation error when they
tried to save (yucky). This wasn't an issue if the user visited
/profile/personal right away, because then the profile regex
pathname check would pass.

Anyway, this logic shouldn't really be in <App /> anyway, so I've moved
it into the relevant component (<PersonalTab />). This should ensure
that any time we render the personal tab we perform the same actions (to
dispatch the setProgram action and ensure the user doesn't see a
needless validation error).

closes #2347
pr #2352