Attach user to coupon (#2392)
* Redeem coupons with `coupon` query param

When the user visits the dashboard with a `coupon` query parameter in the URL,
we will make a POST request to try to attach the user to the coupon specified
by the given coupon code. If successful, we display a confirmation dialog with
the details of the coupon. If unsuccessful, we display a toast notification
informing the user that the coupon is invalid or does not exist.

* correct types

* Improve api tests

* fix lint

* Prevent race condition

* WIP

* Fix test

* Remove unused imports

* Unused imports

* Fix test

* comment

* Only render coupon dialog for coupons