Refactored dashboard API to support getting dashboard for other users
This changes around the dashboard API so that a user may, if they have
the correct permissions, fetch the dashboard for a different user. For
now there is no support in the UI for doing this directly.

A user is permitted to request someone else's dashboard if they are a
'staff' or an 'instructor' on a program which the user who's data they
are requesting is enrolled in.

This also makes a change to the dashboard API where, if the user
making the request is different than the user who's data is being
request, we skip the Edx API refreshing step. This should hopefully make
the API more performant for staff users.

closes #2531
pr #2642