Fixed 'ready to schedule' display for FinalExamCard
This fixes a visual bug in the FinalExamCard. We want to display a list
of all the courses for which the user is currently authorized to take an
exam. Previously we were just showing the program title.

This involves some frontend changes, and a backend change to add an
'can_schedule_exam' boolean field to the courses data that the
dashboard API returns.

closes #2534
pr #2566