Upgraded a few JS dependencies ⬆🆙
When I ran `yarn outdated` today I 😱

So this is the result of running `yarn upgrade` and also specifically
updating 12 dependencies to `@latest`. We have some more dependencies
that are outdated (which we should work on updating soon) but it's a bit
tedious to verify that everything works after upgrading something, so I
didn't want to just do them all now. There are some additional details
on the PR.

pr #2683