Refactored profile form container into an HOC
This refactors the profile form conatiner so that it is implemented as a
HOC (Higher Order Component) instead of a React class. We were
inheriting from this class in a few places before. Now, instead, we wrap
the relevant components with the HOC.

This should make it much more straightforward to change those
components. Yay!

pr #2708