Implemented new navigation design
This changes things around so that the sidebar nav is now used on
desktop as well. This also entails getting rid of the user menu we used
to have (up in the upper right), adding a few more links to the sidebar,
and moving the program selector over to where the user menu used to be.

closes #2564
pr #2836