Implemented Redux REST wrapper
This implements a clean, unified Redux wrapper for use with our REST
APIs. This lets us define little objects (of the Endpoint type) which
define all we need to build a reducer and actions for hitting an API.
Hopefully as we switch APIs over to this we'll be able to see a big
decrease in code size on the client (lots of deletion incoming!).

pr #2941