Implemented AutomaticEmail admin page
This implements an admin-only page at `/automaticemails` which shows the
admin user any automatic emails they have created. Currently, the can
toggle the 'enabled' state of the emails from the page (editing coming
soon!).

In order to implement this I also had to add PATCH handling to the
automatic emails API and I had to also add a foreign key to the
AutomaticEmail table for the User table (so that a User may be
associated with an AutomaticEmail)

closes #3011
pr #3012