Fixed email composiition dialog body loading
This fixes a bug that was introduced in #3145 (mail recipient
variables). In that PR we changed some aspects of how we are managing
the draft.js `EditorState` object, and, in particular, we stopped using
`this.state.editorState === null` to indicate an 'empty' state. Instead,
we started setting it to `EditorState.createEmpty()`, but we did not
update a check (with `R.isNil`) which decided whether a new value needed
to be loaded for the body. This fixes that issue, by getting the
`ContentState` object and calling the `hasText` method on it to
differentiate an empty from a hydrated state.

closes #3181
pr #3185