Fixed automatic email editing
In pr #3238 a bug was introduced into the Automatic Email page, where
the function to PATCH an existing automatic email wasn't being correctly
passed the email ID (which is necessary to get the URL for the email).
This fixes that issue, and also fixes a similar issue that was
preventing the toggling from working, which was introduced in the same
PR. The latter issue was because the mail serializer was being passed
the whole `query` object, and it didn't know what to do with it.

closes #3350
closes #3351