Implement 'recent posts' display on dashboard
This improves the existing 'discussions' card to implement the mocks, by
hitting the OD backend to fetch the frontpage and then rendering it.
Along with that I had to implement a few things to handle hitting OD
with the JWT credential, so we'll have that going forward (lib/auth.js).

closes #3476
pr #3551