fix financial aid skip UI bug
The bug was due to our use of the `fetchJSONWithCSRF` function for
skipping financial aid. The `fetchJSONWithCSRF` function assumes that
the response has a JSON payload, and it tries to parse the payload. If
it doesn't parse correctly it throws an error (well, it returns a
rejected Promise).

The `skipFinancialAid` endpoint doesn't return a JSON payload when it is
`PATCH`'ed, so we were trying to parse JSON where there was none to be
parsed (in fact the endpoint just returns an empty response body).
Anyway, this meant that basically after the request came back as a 200,
but before the 'success' action could be dispatched to the redux store,
we were throwing an error in the `fetchJSONWithCSRF` function that was
causing the failure-case toast message to pop-up. This was a totally
front-end bug, so the request was still going through and skipping the
financial aid, but it wasn't at all obvious to the user what had
happened until they refreshed the page.

So, to fix it I just changed the API function to use `fetchWithCSRF`
instead, which doesn't assume the body of the request contains anything
in particular, and will only reject if the HTTP code is an error code.
This should fix the error in production. I also did a small bit of
cleanup to the relevant function in the `DashboardPage` component.

closes #3803
closes #3034
pr #3805