Add re-enroll button
If the user passed a course, we show them a little 'view certificate'
link. However, in addition to this we'd like to give them the option to
re-enroll in the course. This adds that option.

I discovered that the backend code for the dashboard API was set up to
not return future course runs, in the case where the user already has a
passing final grade on a past run. So, in order to get around that, I
needed to make a small change to the dashboard API code, so that even if
the user passes it will still return a future enrollable run (so we have
something on the frontend in which the user can enroll).

closes #3780
pr #3901