Fix incorrect profile redirect behavior
This fixes a redirect issue we were having, where it was possible for
the user to navigate to the dashboard even if their profile wasn't
completed yet (we assume in a lot of places that this can't happen, so
it breaks stuff if a user does this).

The issue with the existing implementation was that it didn't prevent
the user from navigating away from the profile page after the first
load. So they could go to the profile, click to open the side nav, and
then navigate to the dashboard, and the app wouldn't stop them. Then
they could trigger various API actions, some of which assumed that their
profile would already be complete 😭

I changed it so that the check is done on every prop change on the
top-level App component. So, if the user is on the profile page and then
clicks over to the dashboard, they should be redirected back to the
profile page (and they should see a little toast message).

I also realized that we already perform a redirect on the backend if the
profile is not yet done. There was some code on App.js that was dealing
with the case where the user visits the dashboard first and then gets
redirected to the profile page, but that actually can't happen b/c we
already redirect to the profile page on the backend. So, I decided to
remove that code, both because it was superfluous and b/c keeping it
around complicated things.

closes #3899
pr #3900