Increased the buffer size in uWSGI to address wagtail errors (#3887)
uWSGI allocates a default buffer size of 4096 for request headers. When using certain functionality in Wagtail, the header is much larger than that buffer size so we raise a 502 error due to uWSGI failing to process the request. Increasing the buffer size allows for larger header values, preventing the failure.