Fix learner search page email send error
This PR makes it so that sending a search result email doesn't result in
a failure redux action being dispatched. This was happening because the
frontend was expecting a valid JSON payload in the response, but the
backend wasn't keeping up its end of the bargain. So, I just made a
small change so that the backend will return an empty object instead of
nothing.

closes #3851
pr #3926