fix issue on /profile/personal
this fixes an issue we noticed today. George found the issue, but we
don't know why it's an issue!

Anyhow, for some reason having an import into util/util.js from a
component doesn't seem to work.  This moves a component that was in
util/util.js out to the SendGradesDialog file.

pr #4578